Added a groovy track

+++ MusicPlayer.rb
@@ -6,26 +6,38 @@ class MusicPlayer
 	def initialize
 		aplay = `aplay --version`
 		if /^aplay: version(.+)/ =~ aplay
-			:player = true
+			@player = true
 		end
 	end
 
 	def is_valid?
-		@player == nil
+		@player
 	end
 
 	def play_music
-		@play_thr = Thread.new do
-			loop do
-				@player_pid = spawn "aplay Assets/music_loop.wav -q"
-				Process.wait @player_pid
+		unless @play_thr
+			@play_thr = Thread.new do
+				loop do
+					@player_pid = spawn "aplay Assets/music_loop.wav -q"
+					Process.wait @player_pid
+				end
 			end
 		end
 	end
 
 	def stop_music
-		Process.kill("HUP", @player_pid)
-		@play_thr.kill
+		if (@player_pid || @play_thr)
+			Process.kill("HUP", @player_pid)
+			@play_thr.kill
+			@play_thr = nil
+		end
 	end
 
+	def toggle_music
+		if @play_thr 
+			stop_music
+		else
+			play_music
+		end
+	end
 end
